+++ crates/miden-lib/src/transaction/mod.rs
@@ -22,7 +22,7 @@ mod events;
 pub use events::TransactionEvent;
 
 mod inputs;
-pub use inputs::TransactionAdviceInputs;
+pub use inputs::{TransactionAdviceInputs, TransactionAdviceMapMismatch};
 
 mod outputs;
 pub use outputs::{
@@ -116,7 +116,7 @@ impl TransactionKernel {
         tx_inputs: &TransactionInputs,
         tx_args: &TransactionArgs,
         init_advice_inputs: Option<AdviceInputs>,
-    ) -> (StackInputs, TransactionAdviceInputs) {
+    ) -> Result<(StackInputs, TransactionAdviceInputs), TransactionAdviceMapMismatch> {
         let account = tx_inputs.account();
 
         let stack_inputs = TransactionKernel::build_input_stack(
@@ -127,12 +127,12 @@ impl TransactionKernel {
             tx_inputs.block_header().block_num(),
         );
 
-        let mut tx_advice_inputs = TransactionAdviceInputs::new(tx_inputs, tx_args);
+        let mut tx_advice_inputs = TransactionAdviceInputs::new(tx_inputs, tx_args)?;
         if let Some(init_advice_inputs) = init_advice_inputs {
             tx_advice_inputs.extend(init_advice_inputs);
         }
 
-        (stack_inputs, tx_advice_inputs)
+        Ok((stack_inputs, tx_advice_inputs))
     }
 
     // ASSEMBLER CONSTRUCTOR