+++ crates/miden-tx/src/errors/mod.rs
@@ -1,6 +1,7 @@
 use alloc::{boxed::Box, string::String};
 use core::error::Error;
 
+use miden_lib::transaction::TransactionAdviceMapMismatch;
 use miden_objects::{
     AccountError, Felt, ProvenTransactionError, TransactionInputError, TransactionOutputError,
     Word, account::AccountId, assembly::diagnostics::reporting::PrintDiagnostic,
@@ -16,6 +17,8 @@ use vm_processor::ExecutionError;
 
 #[derive(Debug, Error)]
 pub enum TransactionExecutorError {
+    #[error("the advice map contains conflicting map entries")]
+    ConflictingAdviceMapEntry(#[source] TransactionAdviceMapMismatch),
     #[error("failed to fetch transaction inputs from the data store")]
     FetchTransactionInputsFailed(#[source] DataStoreError),
     #[error("foreign account inputs for ID {0} are not anchored on reference block")]
@@ -79,6 +82,8 @@ pub enum TransactionProverError {
     TransactionOutputConstructionFailed(#[source] TransactionOutputError),
     #[error("failed to build proven transaction")]
     ProvenTransactionBuildFailed(#[source] ProvenTransactionError),
+    #[error("the advice map contains conflicting map entries")]
+    ConflictingAdviceMapEntry(#[source] TransactionAdviceMapMismatch),
     // Print the diagnostic directly instead of returning the source error. In the source error
     // case, the diagnostic is lost if the execution error is not explicitly unwrapped.
     #[error("failed to execute transaction kernel program:\n{}", PrintDiagnostic::new(.0))]