+++ crates/miden-tx/src/executor/exec_host.rs
@@ -1,9 +1,4 @@
-use alloc::{
-    boxed::Box,
-    collections::{BTreeMap, BTreeSet},
-    sync::Arc,
-    vec::Vec,
-};
+use alloc::{boxed::Box, collections::BTreeMap, sync::Arc, vec::Vec};
 
 use miden_lib::{errors::TransactionKernelError, transaction::TransactionEvent};
 use miden_objects::{
@@ -12,13 +7,12 @@ use miden_objects::{
     transaction::{InputNote, InputNotes, OutputNote},
 };
 use vm_processor::{
-    AdviceInputs, BaseHost, ErrorContext, ExecutionError, MastForest, MastForestStore,
-    ProcessState, SyncHost,
+    BaseHost, ErrorContext, ExecutionError, MastForest, MastForestStore, ProcessState, SyncHost,
 };
 
 use crate::{
+    AccountProcedureIndexMap,
     auth::{SigningInputs, TransactionAuthenticator},
-    errors::TransactionHostError,
     host::{ScriptMastForestStore, TransactionBaseHost, TransactionProgress},
 };
 
@@ -61,26 +55,24 @@ where
     pub fn new(
         account: &PartialAccount,
         input_notes: InputNotes<InputNote>,
-        advice_inputs: &mut AdviceInputs,
         mast_store: &'store STORE,
         scripts_mast_store: ScriptMastForestStore,
+        acct_procedure_index_map: AccountProcedureIndexMap,
         authenticator: Option<&'auth AUTH>,
-        foreign_account_code_commitments: BTreeSet<Word>,
-    ) -> Result<Self, TransactionHostError> {
+    ) -> Self {
         let base_host = TransactionBaseHost::new(
             account,
             input_notes,
-            advice_inputs,
             mast_store,
             scripts_mast_store,
-            foreign_account_code_commitments,
-        )?;
+            acct_procedure_index_map,
+        );
 
-        Ok(Self {
+        Self {
             base_host,
             authenticator,
             generated_signatures: BTreeMap::new(),
-        })
+        }
     }
 
     // PUBLIC ACCESSORS