+++ crates/miden-tx/src/prover/mod.rs
@@ -1,6 +1,6 @@
 #[cfg(feature = "async")]
 use alloc::boxed::Box;
-use alloc::{collections::BTreeSet, sync::Arc, vec::Vec};
+use alloc::{sync::Arc, vec::Vec};
 
 use miden_lib::transaction::TransactionKernel;
 use miden_objects::{
@@ -10,11 +10,10 @@ use miden_objects::{
 };
 pub use miden_prover::ProvingOptions;
 use miden_prover::prove;
-use vm_processor::Word;
 use winter_maybe_async::*;
 
 use super::TransactionProverError;
-use crate::host::ScriptMastForestStore;
+use crate::host::{AccountProcedureIndexMap, ScriptMastForestStore};
 
 mod prover_host;
 pub use prover_host::TransactionProverHost;
@@ -88,28 +87,35 @@ impl TransactionProver for LocalTransactionProver {
 
         // execute and prove
         let (stack_inputs, advice_inputs) =
-            TransactionKernel::prepare_inputs(&tx_inputs, &tx_args, Some(advice_witness));
-        let mut advice_inputs = advice_inputs.into_advice_inputs();
+            TransactionKernel::prepare_inputs(&tx_inputs, &tx_args, Some(advice_witness))
+                .map_err(TransactionProverError::ConflictingAdviceMapEntry)?;
 
         // load the store with account/note/tx_script MASTs
         self.mast_store.load_account_code(account.code());
-
-        let account_code_commitments: BTreeSet<Word> = tx_args.foreign_account_code_commitments();
 
         let script_mast_store = ScriptMastForestStore::new(
             tx_args.tx_script(),
             input_notes.iter().map(|n| n.note().script()),
         );
 
-        let mut host = TransactionProverHost::new(
-            &account.into(),
-            input_notes.clone(),
-            &mut advice_inputs,
-            self.mast_store.as_ref(),
-            script_mast_store,
-            account_code_commitments,
-        )
-        .map_err(TransactionProverError::TransactionHostCreationFailed)?;
+        let mut host = {
+            let acct_procedure_index_map = AccountProcedureIndexMap::from_transaction_params(
+                &tx_inputs,
+                &tx_args,
+                &advice_inputs,
+            )
+            .map_err(TransactionProverError::TransactionHostCreationFailed)?;
+
+            TransactionProverHost::new(
+                &account.into(),
+                input_notes.clone(),
+                self.mast_store.as_ref(),
+                script_mast_store,
+                acct_procedure_index_map,
+            )
+        };
+
+        let advice_inputs = advice_inputs.into_advice_inputs();
 
         // For the prover, we assume that the transaction witness was successfully executed and so
         // there is no need to provide the actual source manager, as it is only used to improve